satellite_env/server/requirements.txt
openenv-core>=0.2.2
fastapi>=0.115.0
uvicorn>=0.24.0
pybullet
pydantic>=2.0.0
numpy
openai>=1.0.0
httpx>=0.24.0